Fix sorting events on the client
(k1 < k2) should have been (k1 < k2 ? -1 : 1), but Array.sort() has been
stable for a long time. We don't compare keys on the server side, so
just remove this on the client side as well.

Also, change tiebreak in the leaderboard as suggested in[1]. Instead of
comparing user IDs, which are arbitrary, compare the times that they
found the last set.

Fixes #46

[1] https://github.com/ekzhang/setwithfriends/issues/153